+++ pyscaffold/shell.py
@@ -87,3 +87,7 @@ git = get_git_cmd()
 
 #: Command for django-admin.py
 django_admin = ShellCommand("django-admin.py")
+
+# Command for vagrant
+vagrant = ShellCommand("vagrant")
+