+++ CHANGES.rst
@@ -2,13 +2,15 @@
 Release Notes
 =============
 
-Version 2.5.6, 2016-??-??
+Version 2.5.6, 2016-04-25
 =========================
 
 - Prefix error message with ERROR:
 - Suffix of untagged commits changed from {version}-{hash} to {version}-n{hash}
 - Check if package identifier is valid
 - Added log level command line flags to the skeleton
+- Updated pbr to version 1.9.1
+- Updated setuptools_scm to version 1.11.0
 
 Version 2.5.5, 2016-02-26
 =========================